coxme.object {coxme}	R Documentation

Coxme regression output object

Description

This contains further description of the output object created by a coxme call. Most components can be accessed with extractor functions, which is the safer route since details of the object will likely change over time.

Details

The structure of each element of the random effects coefficients (obtained with ranef) and variances (VarCorr) depend on the variance functions, i.e., the functions used in the varlist argument. Since users can write their own variance functions this format can never be completely known.

Value

coefficients
the coefficients of the fixed effects. Use the fixef function to extract them.

frail
the coefficients of the random effects. Use the ranef function to extract them. These are always stored as a list with one member per random effect; each parenthesised term in the model is a random effect. In a linear mixed effects model the fixed effects and the variances of the random effects can be obtained without explicitly computing the coefficients of the random effects, the latter are called BLUP estimates and are computed later if at all. This is not the case for a Cox model, there the random effect coefficients are a required part of the iteration process and so are always present in the final model.

vcoef
the variances of the random effects. Use the VarCorr function to extract them. These are always stored as a list with one member per random effect.

variance
the variance-covariance matrix of the coefficient vector, including both fixed and random terms. The random effects are listed first. This will often be a sparse matrix. The vcov function will extract the fixed effects portion, which is always dense.

loglik
the log-likelihood vector from the fit. The first element is the loglik at the initial values, the second is the integrated partial likelihood at the solution (IPL), the third is the penalized partial likelihood at the solution(PPL).

df
degrees of freedom for the IPL and the PPL solutions.

hmat
sparse Cholesky factorization of the information matrix.

iter
outer and inner iterations performed. For each trial value of the variance parameters an Cox model partial likelihood must be solved; the outer iterations is the reported number from the optim() routine which handles the variance parameters, the inner iterations is the cumulative number of partial likelihood iterations.

control
a copy of the coxph.control parameters used in the fit.

ties
the computational method used for ties.

u
the vector of first derivatives of the PPL, at the solution.

means,scale
means and scale for each predictor, used internally to scale the problem.

linear.predictor
the vector of linear predictors.

n
vector containing the number of events and the number of observations in the fitting data set.

terms
the terms object from the fixed effects of the model formula. Access using the terms function.

formulaList
the fixed and random portions of the formula, separated

na.action
the missing value attributes of the data, if any

x,y,model
optional: the x matrix, response, for model frame. These depend on the corresponding arguments in the call.

call
a copy of the call to the routine

[Package coxme version 2.2-17 Index]