coxmeFull {coxme}	R Documentation

Variance family function for coxme fits.

Description

This function sets up the default variance family information for a mixed effects survival model fit with coxme.

Usage

coxmeFull(collapse = FALSE)

Arguments

collapse
Form for fitting a nested effect, either standard or collapsed. The latter appears to be more numerically stable (still under research).

Details

Coxme variance families create a list with three functions: initialize, generate, and wrapup, that determine how the variance structure of a fit is modeled.

Value

an object of class coxvar.

Author(s)

Terry Therneau

See Also

coxme

[Package coxme version 2.2-17 Index]