coxmeMlist {coxme}	R Documentation

Coxme variance function

Description

This variance function accepts a list of matrices, which define a correlation structure for a coxme fit.

Usage

coxmeMlist(varlist, rescale = FALSE, pdcheck = TRUE, positive = TRUE)

Arguments

varlist
a list containing one or more matrix or bdsmatrix objects.

rescale
if TRUE, each input matrix is rescaled to have a diagonal of 1. (Kinship matrices for instance are often generated with a diagonal of .5 and would be multiplied by 2).

pdcheck
check each matrix to ensure that it is positive definite

positive
constrain coefficients to be positive. This may also be a vector of the same length as varlist

Details

If two matrices A and B were given, this fits the variance structure V = s1 A + s2 B, where the variances s1 and s2 are parameters that will be optimized by coxme, treating A and B as fixed.

Value

a coxme variance family object, used by coxme in the fitting process.

Author(s)

Terry Therneau

See Also

coxme

[Package coxme version 2.2-17 Index]