eortc {coxme}	R Documentation

Simulated data set based on an EORTC trial

Description

This is a simulated surival data set for investigating random center effects. To make it realistic, the number of centers and their sizes is based on an EORTC cancer trial.

Usage

data(eortc)

Format

A data frame with 2323 observations on the following 4 variables.

y

survival time

uncens

0=alive, 1=dead

center

enrolling center, a number from 1 to 37

trt

treatment arm, 0 or 1

Details

This is used in the test suite for the code.

Source

PhD thesis work of Jose Cortinas Abrahantes

References

Cortinas Abrahantes, Jose; Burzykowski, Tomasz (2002), A version of the EM algorithm for proportional hazards models with random effects , Published in: Lecture Notes of the ICB Seminars. p. 15-20

Examples

data(eortc)
coxme(Surv(y, uncens) ~ trt + (trt| center) + strata(center), eortc)

[Package coxme version 2.2-17 Index]