expand.nested {coxme}	R Documentation

Expand nested factors

Description

Expand out the data frame for a nested factor such as (1| a/b). This is used by the variance function routines of coxme.

Usage

expand.nested(x)

Arguments

x
A data frame containing the nesting variables

Details

The initialize function of a coxme variance family is passed, as one of its arguments, a data frame G containing the grouping variables, each of which is a factor.. Assume a nested factor (1| a/b) in the model formula and a data set whose first few rows are:

    a    b
    1    1
    1    2
    2    1
  

The function will replace the second column with a variable named a/b and values of 1/1, 1/2, 2/1, etc.

Value

an updated data frame

Author(s)

Terry Therneau

See Also

coxme, coxmeMlist

[Package coxme version 2.2-17 Index]