fixef.lmekin {coxme}	R Documentation

Extraction functions for Lmekin

Description

Extract the fixed effects, random effects, variance of the fixed effects, or variance of the random effects from a linear mixed effects model fit with lmekin.

Usage

## S3 method for class 'lmekin'
fixef(object, ...)
## S3 method for class 'lmekin'
ranef(object, ...)
## S3 method for class 'lmekin'
vcov(object, ...)
## S3 method for class 'lmekin'
VarCorr(x, ...)
## S3 method for class 'lmekin'
logLik(object, ...)

Arguments

object
an object inheriting from class lmekin representing the result of a mixed effects model.

x
an object inheriting from class lmekin representing the result of a mixed effects model.

...
some methods for this generic require additional arguments. None are used in this method.

Details

For the random effects model y = X beta + Zb + e, let s^2 be the variance of the error term e. Let A = sP be the variance of the random effects b. There is a computational advantage to solving the problem in terms of P instead of A, and that is what is stored in the returned lmekin object. The VarCorr function returns elements of P; the print and summary functions report values of A. Pinhiero and Bates call P the precision factor.

Value

the fixed effects are a vector and vcov returns their variance/covariance matrix. The random effects are a list with one element for each random effect. The ranef component contains the coefficients and VarCorr the estimated variance/covariance matrix. The logLik method returns the loglikelihood along with its degrees of freedom.

Author(s)

Terry Therneau

References

J Pinheiro and D Bates, Mixed-effects models in S and S-Plus. Springer, 2000.

See Also

lmekin, random.effects, fixed.effects, link{vcov}, VarCorr

Examples

data(ergoStool, package="nlme")  # use a data set from nlme
efit <-  lmekin(effort ~ Type + (1|Subject), ergoStool)
ranef(efit)

[Package coxme version 2.2-17 Index]