lmekin.object {coxme}	R Documentation

lmekin object

Description

This class of object is returned by the lmekin function to represent a fittd mixed effect linear model. Objects of this class currently have methods for print and residuals.

Value

A list with the folling components:

coefficients
a list with components fixed and random; the first will be NULL for a model with no fixed effects. The random component is itself a list, with an element for each random effect.

var
the variance matrix of the fixed effects

vcoef
the parameters of the variance matrix of the random effects.

residuals
vector of residuals from the fit

method
either "ML" or "REML"

loglik
the log-likelihood for the fitted model

sigma
the estimated residual error

n
number of observations used

call
a copy of the call

na.action
this will be present if any observations were removed due to missing values

Author(s)

Terry Therneau

See Also

lmekin, coxmeFull, coxmeMlist

[Package coxme version 2.2-17 Index]