logLik.coxme {coxme}	R Documentation

The logLik method for coxme objects

Description

logLik is most commonly used for a model fitted by maximum likelihood, and some uses, e.g. by AIC. This method allows generic functions to easily extract the log-likelhood of a coxme model.

Usage

## S3 method for class 'coxme'
logLik(object, type = c("penalized", "integrated"), ...)

Arguments

object
a fitted coxme model

type
which of the two types of partial likelihood to extract

...
used by other methods

Details

The likelihood for a mixed effects Cox model can be viewed in two ways: the ordinarly partial likelihood, where the random effects act only as a penalty or constraint, or a partial likelihood where the random effect has been integrated out. Both are valid.

Value

Returns an object of class logLik.

See Also

logLik

[Package coxme version 2.2-17 Index]