print.coxme {coxme}	R Documentation

Print method for a coxme fit.

Description

Print out the result of a coxme fit.

Usage

## S3 method for class 'coxme'
print(x, rcoef=FALSE, digits = options()$digits, ...)
## S3 method for class 'coxme'
summary(object, ...)

Arguments

x
an object of class coxme, from the fit of a mixed effects survival model.

rcoef
print the random (penalized) coefficients, as well as the fixed ones.

digits
number of significant digits to print

object
an object of class coxme, from the fit of a mixed effects survival model.

...
optional arguments

Note

The summary function is currently identical to the print function. The default in R is for print() to be the short form printout and summary() the long form.

Author(s)

Terry Therneau

See Also

coxme

[Package coxme version 2.2-17 Index]